+++ src/main/kotlin/com/calmwolfs/bedwar/config/features/DevConfig.java
@@ -1,7 +1,6 @@
 package com.calmwolfs.bedwar.config.features;
 
 import com.google.gson.annotations.Expose;
-import io.github.moulberry.moulconfig.annotations.ConfigEditorBoolean;
 import io.github.moulberry.moulconfig.annotations.ConfigEditorText;
 import io.github.moulberry.moulconfig.annotations.ConfigOption;
 
@@ -10,10 +9,4 @@ public class DevConfig {
     @ConfigOption(name = "Api Key", desc = "Put your api key here. §cDo not use this unless you are testing something")
     @ConfigEditorText
     public String apiKey = "";
-
-    @Expose
-    @ConfigOption(name = "Repo Auto Update", desc = "Update the repository on every startup.\n" +
-            "§eOnly disable this if you know what you are doing!")
-    @ConfigEditorBoolean
-    public boolean repoAutoUpdate = true;
 }